% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{clearvariables}
\alias{clearvariables}
\title{Eliminar variables}
\usage{
clearvariables()
}
\description{
La funcion clearvariables() elimina todas las variables del entorno R
}
\examples{
clearvariables()
}
