% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{rstring}
\alias{rstring}
\title{Replicar string}
\usage{
rstring(s, num)
}
\arguments{
\item{s}{cadena a replicar}

\item{nun}{numero de veces a replicar la cadena s}
}
\description{
La funcion rstring() devuelve un vector de num posiciones con el valor de la cadena s
}
\examples{
rstring("mi cadena", 10)
}
